Add working flak for non wave file. Only works on small files though, large files result in  a seg fault.

+++ FLAK/FLAK.h
@@ -42,7 +42,7 @@ struct ErrorWrapper {
     I e1[frame_sample_size_const];
     I e2[frame_sample_size_const];
     I e3[frame_sample_size_const];
-    long sums[4];
+    long long sums[4];
 };
 
 class FLAKCOMP {
